Add System and None (JVM Default) themes.
The current 2 editor color schemes are used depending on whether a LAF is detected to be dark ot light.
This is done by drawing an empty JPanel into a dummy image and getting the grayscale pixel brightness.
Move setting theme a bit earlier so it doesn't throw exceptions in some cases when using system theme.
Change the order of setting swing LAF and applying theme so that apply() can get the right colors.