Do not abort decompile command on decompiler errors
As explained in PR #289, the decompilers may fail decompiling some
vanilla Minecraft classes. Although it is my opinion that these errors
should be fixed in the decompilers themselves, currently they render the
decompile CLI command useless, as no output is produced whatsoever.

The PR that was mentioned previously addressed this situation for the
GUI part of Enigma, but not for the CLI commands, which were left
untouched.

These changes modify the decompiler error strategy used by the
aforementioned CLI command to trace the errors, but continue decompiling
classes, so at least a partial output is generated. This way users will
be happy that the Enigma decompile command works, and developers will
still be motivated to fix the underlying issues.

Signed-off-by: AlexTMjugador <AlexTMjugador@users.noreply.github.com>