Allow disabling editing part of the mappings
For now, this is done over CLI options, it cannot be changed through the
GUI. The new CLI options are:

    -C, --no-edit-classes     Disable editing class names
    -D, --no-edit-javadocs    Disable editing Javadocs
    -E, --no-edit-all         Disable editing everything
    -F, --no-edit-fields      Disable editing field names
    -M, --no-edit-methods     Disable editing method names
    -P, --no-edit-parameters  Disable editing parameter names
    -c, --edit-classes        Enable editing class names
    -d, --edit-javadocs       Enable editing Javadocs
    -e, --edit-all            Enable editing everything
    --edit-locals             Enable editing local variable names
    -f, --edit-fields         Enable editing field names
    -m, --edit-methods        Enable editing method names
    --no-edit-locals          Disable editing local variable names
    -p, --edit-parameters     Enable editing parameter names

They can be combined together and are applied in order of appearance in the
argument list from left to right, and everything is editable by default,
as before. For example,

    --no-edit-all --edit-javadocs --edit-parameters

(or -Edp for short) will allow you to edit only Javadocs and parameter
names. --(no-)edit-locals is only provided for completeness and has no
effect right now because Enigma does not allow for editing locals other
than parameters.