Update Gradle and fix buildscripts (#557)
* Fix Gradle configuration cache compat with generateResources task

The task now also declares its inputs properly, making it support
up-to-date checks.

* Update Gradle and Shadow

Shadow 9 is incompatible with including the flatlaf natives,
so I updated to the last 8.x version which also works with Gradle 9.

The Shadow plugin automatically publishes the shadow jar, so I removed
the overlapping publications which overrode each other.

* Fix more Gradle 10 deprecations

* Enable configuration cache by default

* Fix missing task dependency of :enigma-cli:test